+++ RELEASE_NOTES.md
@@ -15,7 +15,8 @@ This serves two purposes:
 ### Changed
 - internal: The HydeKernel has been refactored to move related logic to service classes. This does not change the end usage as the Hyde facade still works the same
 - `DiscoveryService::getSourceFileListForModel()` now throws an exception instead of returning false when given an invalid model class
-- `DiscoveryService::getFilePathForModelClassFiles` helper was renamed to `DiscoveryService::getModelSourceDirectory`
+- `DiscoveryService::getFilePathForModelClassFiles` method was renamed to `DiscoveryService::getModelSourceDirectory`
+- `DiscoveryService::getFileExtensionForModelFiles` method was renamed to `DiscoveryService::getModelFileExtension`
 
 ### Deprecated
 - for soon-to-be removed features.